Add files via upload
To this point in my project, I have been able to create a program that will add users, books, and exit. 
Also in the project, I have been able to create a navigation menu whereby everything in the system can be managed or rather else can be done via the menu that entails them all.